+++ src/Dialog/index.js
@@ -39,13 +39,6 @@ export default class Dialog extends Popup {
        * @type {HTMLElement|NodeList|Array}
        */
       content: [],
-
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
     };
 
     // Merge remaining options with defaults and save all as instance properties.
@@ -103,7 +96,9 @@ export default class Dialog extends Popup {
     );
 
     // Fire the init event.
-    this.dispatch('init', { instance: this });
+    if (! this._suppressDispatch.includes('init')) {
+      this.dispatch('init', { instance: this });
+    }
   }
 
   /**
@@ -218,7 +213,9 @@ export default class Dialog extends Popup {
     this.target.removeEventListener('keydown', this.targetHandleKeydown);
     document.body.removeEventListener('keydown', this.handleKeydownEsc);
 
-    /* Run {destroyCallback} */
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    if (! this._suppressDispatch.includes('destroy')) {
+      this.dispatch('destroy', { element: this.element });
+    }
   }
 }