+++ src/Listbox/index.js
@@ -31,22 +31,8 @@ export default class ListBox extends Popup {
     this.controller = controller;
     this.target = super.constructor.getTargetElement(controller);
 
-    /**
-     * Options shape.
-     *
-     * @type {object}
-     */
-    const defaultOptions = {
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
-    };
-
-    // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, defaultOptions, options);
+    // Merge options as instance properties.
+    Object.assign(this, options);
 
     // Bind class methods.
     this.preventWindowScroll = this.preventWindowScroll.bind(this);
@@ -132,7 +118,9 @@ export default class ListBox extends Popup {
     window.addEventListener('keydown', this.preventWindowScroll);
 
     // Fire the init event.
-    this.dispatch('init', { instance: this });
+    if (! this._suppressDispatch.includes('init')) {
+      this.dispatch('init', { instance: this });
+    }
   }
 
   /**
@@ -395,7 +383,9 @@ export default class ListBox extends Popup {
     this.target.removeEventListener('blur', this.targetHandleBlur);
     window.removeEventListener('keydown', this.preventWindowScroll);
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    if (! this._suppressDispatch.includes('destroy')) {
+      this.dispatch('destroy', { element: this.element });
+    }
   }
 }