+++ src/Menu/index.js
@@ -71,13 +71,6 @@ export default class Menu extends AriaComponent {
        * @type {string}
        */
       itemMatches: '*',
-
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
     };
 
     // Merge remaining options with defaults and save all as instance properties.
@@ -180,7 +173,7 @@ export default class Menu extends AriaComponent {
       if (this.collapse && link.hasAttribute('target')) {
         const disclosure = new Disclosure(
           link,
-          { _suppressDispatch: ['init'] }
+          { _suppressDispatch: ['init', 'destroy'] }
         );
 
         this.disclosures.push(disclosure);
@@ -191,7 +184,7 @@ export default class Menu extends AriaComponent {
         // Instantiate sub-Menus.
         const subList = new Menu(
           siblingList,
-          { _suppressDispatch: ['init'] }
+          { _suppressDispatch: ['init', 'destroy'] }
         );
 
         // Save the list's previous sibling.
@@ -378,7 +371,9 @@ export default class Menu extends AriaComponent {
       disclosure.destroy();
     });
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    if (! this._suppressDispatch.includes('destroy')) {
+      this.dispatch('destroy', { element: this.element });
+    }
   }
 }