+++ src/MenuBar/index.js
@@ -69,13 +69,6 @@ export default class MenuBar extends AriaComponent {
        * @type {string}
        */
       itemMatches: '*',
-
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
     };
 
     // Merge remaining options with defaults and save all as instance properties.
@@ -220,7 +213,7 @@ export default class MenuBar extends AriaComponent {
         list,
         {
           itemMatches: this.itemMatches,
-          _suppressDispatch: ['init'],
+          _suppressDispatch: ['init', 'destroy'],
         }
       );
       target.addEventListener('keydown', this.menuItemHandleKeydown);
@@ -250,7 +243,9 @@ export default class MenuBar extends AriaComponent {
     rovingTabIndex(this.menuBarItems, this.firstItem);
 
     // Fire the init event.
-    this.dispatch('init', { instance: this });
+    if (! this._suppressDispatch.includes('init')) {
+      this.dispatch('init', { instance: this });
+    }
   }
 
   /**
@@ -449,7 +444,9 @@ export default class MenuBar extends AriaComponent {
       submenu.destroy();
     });
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    if (! this._suppressDispatch.includes('destroy')) {
+      this.dispatch('destroy', { element: this.element });
+    }
   }
 }