+++ src/Tablist/index.js
@@ -38,22 +38,8 @@ export default class Tablist extends AriaComponent {
 
     this.tabs = tabs;
 
-    /**
-     * Component configuration options.
-     *
-     * @type {object}
-     */
-    const defaultOptions = {
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
-    };
-
-    // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, defaultOptions, options);
+    // Merge options as instance properties.
+    Object.assign(this, options);
 
     // Bind class methods.
     this.panelHandleKeydown = this.panelHandleKeydown.bind(this);
@@ -432,7 +418,9 @@ export default class Tablist extends AriaComponent {
       );
     });
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    if (! this._suppressDispatch.includes('destroy')) {
+      this.dispatch('destroy', { element: this.element });
+    }
   }
 }